fix: move nudgeSameDayDueDate to flashcard-fsrs to avoid DB import in tests
flashcard-data.ts pulls in the database client, so importing from it in
tests requires DATABASE_URL. Moving the pure nudgeSameDayDueDate function
to flashcard-fsrs.ts (which has no DB dependency) lets the seed tests
run in CI without a database connection.
